+++ test/authentication.spec.ts
@@ -1,12 +1,23 @@
 import { expect } from 'chai'
 import { join } from 'path'
 import { setupServer } from './setup'
-import { Preset } from '../src'
+import { Preset, CacheType } from '../src'
 
 describe('Basic authentication', function() {
   const server = setupServer({
-    root: join(__dirname, 'cra'),
-    preset: Preset.CRA,
+    folders: [
+      {
+        path: '/static',
+        cache: CacheType.Immutable,
+        root: join(__dirname, 'cra/static'),
+      },
+      {
+        path: '/',
+        cache: CacheType.Short,
+        root: join(__dirname, 'cra'),
+      },
+    ],
+    index: join(__dirname, 'cra/index.html'),
     username: 'test',
     password: '1234',
   })