+++ test/prefetch.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { join } from 'path'
 import { setupServer } from './setup'
-import { Preset } from '../src'
+import { Preset, CacheType } from '../src'
 
 export function atob(what: string) {
   return Buffer.from(what, 'base64').toString('utf-8')
@@ -10,8 +10,19 @@ export function atob(what: string) {
 describe('Prefetch tag injection', function() {
   describe('CRA', function() {
     const server = setupServer({
-      root: join(__dirname, 'cra'),
-      preset: Preset.CRA,
+      folders: [
+        {
+          path: '/static',
+          cache: CacheType.Immutable,
+          root: join(__dirname, 'cra/static'),
+        },
+        {
+          path: '/',
+          cache: CacheType.Short,
+          root: join(__dirname, 'cra'),
+        },
+      ],
+      index: join(__dirname, 'cra/index.html'),
       prefetch: true,
     })