+++ test/sri.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { join } from 'path'
 import { setupServer } from './setup'
-import { Preset } from '../src'
+import { Preset, CacheType } from '../src'
 import crypto from 'crypto'
 import cheerio from 'cheerio'
 import { AxiosResponse } from 'axios'
@@ -15,8 +15,19 @@ export function hash(what: AxiosResponse<any>) {
 
 describe('SRI tag injection', function() {
   const server = setupServer({
-    root: join(__dirname, 'cra'),
-    preset: Preset.CRA,
+    folders: [
+      {
+        path: '/static',
+        cache: CacheType.Immutable,
+        root: join(__dirname, 'cra/static'),
+      },
+      {
+        path: '/',
+        cache: CacheType.Short,
+        root: join(__dirname, 'cra'),
+      },
+    ],
+    index: join(__dirname, 'cra/index.html'),
     sri: true,
   })