+++ test/utils.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { join } from 'path'
 import { Preset, CacheType } from '../src'
-import { readCli, snakeToCamelCase, readFoldersFromEnv, validateSPAServerConfig } from '../src/util'
+import { snakeToCamelCase, readFoldersFromEnv, validateSPAServerConfig } from '../src/util'
 
 describe('validateSPAServerConfig', function() {
   // it('Should not start with no folder, no root and no preset', async function() {
@@ -14,18 +14,19 @@ describe('validateSPAServerConfig', function() {
   //   ).to.eventually.be.rejected
   // })
 
-  it('Nonexistant index should be forbidden', function() {
-    expect(() =>
+  it('Nonexistant index should be forbidden', async function() {
+    await expect(
       validateSPAServerConfig({
         port: 0,
         silent: true,
         index: join(__dirname, 'index.html'),
+        folders: [],
       })
-    ).to.throw()
+    ).to.eventually.be.rejectedWith(/index/)
   })
 
-  it('Nonexistant folder should be forbidden', function() {
-    expect(() =>
+  it('Nonexistant folder should be forbidden', async function() {
+    await expect(
       validateSPAServerConfig({
         port: 0,
         index: join(__dirname, 'basic', 'index.html'),
@@ -36,77 +37,7 @@ describe('validateSPAServerConfig', function() {
           },
         ],
       })
-    ).to.throw()
-  })
-
-  it('Completely broken config should not work at all', function() {
-    expect(() =>
-      validateSPAServerConfig({
-        port: 0,
-        index: join(__dirname, 'basic', 'index.html'),
-        silent: true,
-      } as any)
-    ).to.throw(/root/)
-  })
-
-  it('Root without preset does not work', function() {
-    expect(() =>
-      validateSPAServerConfig({
-        port: 0,
-        index: join(__dirname, 'basic', 'index.html'),
-        silent: true,
-        root: join(__dirname, 'basic'),
-      } as any)
-    ).to.throw(/root/)
-  })
-
-  it('Root with preset works', function() {
-    expect(() =>
-      validateSPAServerConfig({
-        port: 0,
-        index: join(__dirname, 'basic', 'index.html'),
-        silent: true,
-        root: join(__dirname, 'basic'),
-        preset: Preset.None,
-      } as any)
-    ).to.not.throw(/root/)
-  })
-
-  it('Root with preset and folders is forbidden', function() {
-    expect(() =>
-      validateSPAServerConfig({
-        port: 0,
-        index: join(__dirname, 'basic', 'index.html'),
-        silent: true,
-        root: join(__dirname, 'basic'),
-        preset: Preset.None,
-        folders: [
-          {
-            root: join(__dirname, 'cra'),
-            cache: CacheType.Short,
-          },
-          {
-            path: '/static',
-            root: join(__dirname, 'cra', 'static'),
-            cache: CacheType.Long,
-          },
-        ],
-      } as any)
-    ).to.throw()
-  })
-})
-
-describe('readCli', function() {
-  it('Should read port', async function() {
-    await expect(readCli(['--port', '80', '--root', join(__dirname, 'basic')])).to.deep.include({
-      port: 80,
-    })
-  })
-
-  it('Should read root', async function() {
-    await expect(readCli(['--port', '80', '--root', join(__dirname, 'basic')])).to.deep.include({
-      root: join(__dirname, 'basic'),
-    })
+    ).to.eventually.be.rejectedWith(/folders/)
   })
 })